Update overcommit config
The RuboCop check should be invoked using bundler so that the correct
RuboCop version is used, otherwise we'll get unexpected results.
Also, We can re-enable the AuthorName check because the original
behavior of the check that prompted it to be disabled has been changed
upstream; first and last name are no longer required, anything of non-0
length is sufficient.