Use Docker executor in CI for deploying docs (#137)
* Use Docker executor in CI for deploying docs

Not specifying a machine image defaults to using a Trusty image, which
is deprecated and will soon fail. Assuming this works correctly,
it seems like using a Docker executor is preferable anyway.
(see https://circleci.com/blog/ubuntu-14-16-image-deprecation/)

* Add GitHub's host SSH key to known_hosts

This is necessary because we use a custom checkout step when deploying
docs. CircleCI's built-in `checkout` step does this automatically.