Fix edit_application_command with global scope.
A condition has errors when editing globally scoped command.
Firstly, #type is called on object that might be nil, which results
into:
"Exception: #<NoMethodError: undefined method `type' for nil:NilClass>"
in the logs.

Secondly, fix the method name itself as it should be
edit_global_command, since guild commands are dealt with in previous
branch.