Undo Style/SafeNavigation autocorrect
Normally you would be able to use the safe navigation (&.) to check for missing
values and automatically run `.map` if it's set. However, it appears this
sometimes gets passed `false` instead of nil, which is breaking specs.

The code should probably be refactored, but leaving that as a future item to
reduce the scope of this PR.